Remove ability to convert opencontrols to oscal
I have found that converting opencontrols to oscal cannot be done 1-to-1. It
will be always opinionated conversion.

Most common conversion may be conversion of opencontrol to oscal fedramp ssp. We
don't want to bring dependencies on fedramp to oscalkit. Fedramp is very
opinionated ssps and we want to keep oscalkit versatile.

In order to convert opencontrols to FedRAMP SSPs please consult sister project
of oscalkit:

    https://github.com/gocomply/fedramp