Rename automatically generated files
This achieves consistency, you always search for the same file within a
package to see the models.

This achieves clarity, file named generated_models can be expected to
contain generated_models.

This allows for profile implementation to be in profile.go, not
profile_imp.go (as was done previously).

This allows us to generate more stuff than just models and place it
along the generated_models.go file.

Note: Re-generating models with the latest metaschema.